the initial version of the program was uploaded. The program performs the following tasks:
1. Simulates the roll of a die a specified number of times and calculates the relative frequency of each face.

2. Performs a conditional dice roll simulation based on pre-specified probabilities for each face.

3. Simulates the motion of a celestial body in an astronomical system using a Runge-Kutta method for solving ordinary differential equations.

The code is written in Fortran and utilizes various Fortran intrinsic functions such as random_number and a user-defined ordinary differential equation solver. The results of the simulations are printed to the console and also saved to data files for further analysis.

The program lacks proper modularization and detailed comments explaining each step of the program, and these improvements will be addressed in future commits.